Split up JsonCodec
